Subin

+++ src/controllers/petController.ts
@@ -8,8 +8,13 @@ import { addPet, deletePet, getALlPet, getPet } from '../services/pet.service';
 export const addPetHandler = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const { user_id } = req.params;
-    const { pet_name, pet_gender, pet_species, pet_birth, pet_info, pet_img } = req.body;
 
+    if (req.body.jwtDecoded.userId !== req.params.user_id) {
+      throw new Error('[ 요청 에러 ] 본인이 아닙니다.');
+    }
+
+    const { pet_name, pet_gender, pet_species, pet_birth, pet_info } = req.body;
+    const imgFileRoot = `http://localhost:3000/api/v1/static/${req.file?.filename}`;
     if (
       !user_id ||
       !pet_name ||
@@ -17,7 +22,7 @@ export const addPetHandler = async (req: Request, res: Response, next: NextFunct
       !pet_species ||
       !pet_birth ||
       !pet_info ||
-      !pet_img
+      !imgFileRoot
     )
       throw new Error('[ 요청 에러 ] 모든 필드를 입력해야 합니다.');
 
@@ -28,7 +33,7 @@ export const addPetHandler = async (req: Request, res: Response, next: NextFunct
       pet_species,
       pet_birth,
       pet_info,
-      pet_img,
+      pet_img: imgFileRoot,
     };
 
     const createdPet = await petService.addPet(petData);
@@ -45,10 +50,14 @@ export const addPetHandler = async (req: Request, res: Response, next: NextFunct
 // pet 상세 조회
 export const getPetHandler = async (req: Request, res: Response, next: NextFunction) => {
   try {
-    const pet_id = parseInt(req.params.petId);
+    const pet_id = parseInt(req.params.pet_id);
     if (!pet_id) throw new Error('[ 요청 에러 ] pet_id가 필요합니다.');
 
     const pet = await petService.getPet(pet_id);
+
+    if (req.body.jwtDecoded.userId !== pet.user_id) {
+      throw new Error('[ 요청 에러 ] 본인의 펫이 아닙니다.');
+    }
 
     res.status(200).json({ message: 'pet 상세 조회 성공', data: pet });
   } catch (error: any) {
@@ -61,8 +70,10 @@ export const getPetHandler = async (req: Request, res: Response, next: NextFunct
 // pet 전체 조회
 export const getAllPetHandler = async (req: Request, res: Response, next: NextFunction) => {
   try {
-
-    const {user_id}  = req.body;
+    if (req.body.jwtDecoded.userId !== req.body.user_id) {
+      throw new Error('[ 요청 에러 ] 본인의 펫만 조회 가능합니다.');
+    }
+    const { user_id } = req.body;
     const pets = await petService.getALlPet(user_id);
 
     res.status(200).json({ message: '전체 pets 조회 성공', data: pets });
@@ -73,16 +84,26 @@ export const getAllPetHandler = async (req: Request, res: Response, next: NextFu
   }
 };
 
-// pet 수정
+// pet 수정 >> Todo: req.body.jwtDecoded 값이 출력이 안되는거 수정하기
 export const updatePetHandler = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const pet_id = parseInt(req.params.pet_id);
-    const {user_id}  = req.body;
+
     if (!pet_id) throw new Error('[ 요청 에러 ] pet_id가 필요합니다.');
 
-    const { pet_name, pet_gender, pet_species, pet_birth, pet_info, pet_img } = req.body;
+    const { pet_name, pet_gender, pet_species, pet_birth, pet_info, user_id } = req.body;
 
-    if ( !user_id ||!pet_name || !pet_gender || !pet_species || !pet_birth || !pet_info || !pet_img )
+    const imgFileRoot = `http://localhost:3000/api/v1/static/${req.file?.filename}`;
+
+    if (
+      !user_id ||
+      !pet_name ||
+      !pet_gender ||
+      !pet_species ||
+      !pet_birth ||
+      !pet_info ||
+      !imgFileRoot
+    )
       throw new Error('[ 요청 에러 ] 변경된 값이 없습니다!');
 
     const updatePetData: Pet.updatePetInput = {
@@ -91,9 +112,18 @@ export const updatePetHandler = async (req: Request, res: Response, next: NextFu
       pet_species,
       pet_birth,
       pet_info,
-      pet_img,
+      pet_img: imgFileRoot,
       user_id,
     };
+
+    const findPet = await petService.getPet(pet_id);
+    if (findPet.user_id !== updatePetData.user_id) {
+      throw new Error('[ pet 업데이트 실패 ] 본인의 펫이 아닙니다.');
+    }
+    /*if(req.body.jwtDecoded.userId !==updatePetData.user_id) {
+      throw new Error('[ pet 업데이트 실패 ] 본인의 펫이 아닙니다.');
+    }
+*/
     const pet = await petService.updatePet(pet_id, updatePetData);
 
     res.status(200).json({ message: 'pet 수정 성공', data: pet });
@@ -109,6 +139,11 @@ export const deletePetHandler = async (req: Request, res: Response, next: NextFu
   try {
     const pet_id = parseInt(req.params.pet_id);
     if (!pet_id) throw new Error('[ 요청 에러 ] pet_id가 필요합니다.');
+
+    const pet = await petService.getPet(pet_id);
+    if (req.body.jwtDecoded.userId !== pet.user_id) {
+      throw new Error('[ 요청 에러 ] 본인의 펫이 아닙니다.');
+    }
 
     const petId = await petService.deletePet(pet_id);
